requirements.txt
pandas==2.1.0
opensearch-py-ml
opensearch-py
python-dotenv
requests
deprecated